smiplify kotlin iOS files
