+++ samples/SkiaMultiplatformSample/src/iosMain/kotlin/org/jetbrains/skiko/sample/SwiftHelper.kt
@@ -1,22 +1,7 @@
-import org.jetbrains.skiko.sample.Clocks
-import platform.UIKit.UIViewController
-import kotlinx.cinterop.*
-import org.jetbrains.skiko.*
-import platform.UIKit.*
-import platform.Foundation.*
-import org.jetbrains.skiko.sample.*
+package org.jetbrains.skiko.sample
+
+import org.jetbrains.skiko.SkiaLayer
 
 class SwiftHelper {
     fun getViewController() = getSkikoViewContoller()
 }
-
-fun makeApp(skiaLayer: SkiaLayer) = Clocks(skiaLayer)
-
-fun getSkikoViewContoller(): UIViewController = SkikoViewController(
-    SkikoUIView(
-        SkiaLayer().apply {
-            gesturesToListen = SkikoGestureEventKind.values()
-            skikoView = GenericSkikoView(this, makeApp(this))
-        }
-    )
-)
