+++ samples/SkiaMultiplatformSample/src/iosMain/kotlin/org/jetbrains/skiko/sample/getSkikoViewContoller.kt
@@ -1,0 +1,18 @@
+package org.jetbrains.skiko.sample
+
+import kotlinx.cinterop.*
+import org.jetbrains.skiko.*
+import org.jetbrains.skiko.sample.*
+import platform.Foundation.*
+import platform.UIKit.*
+
+fun makeApp(skiaLayer: SkiaLayer) = Clocks(skiaLayer)
+
+fun getSkikoViewContoller(): UIViewController = SkikoViewController(
+    SkikoUIView(
+        SkiaLayer().apply {
+            gesturesToListen = SkikoGestureEventKind.values()
+            skikoView = GenericSkikoView(this, makeApp(this))
+        }
+    )
+)
